# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Autonomous AI Data Scientist API","description":"Durable analysis workflows. Numeric computations run in Python; the optional LLM supplies narrative reasoning only.","version":"1.0.0"},"paths":{"/api/health":{"get":{"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/datasets":{"get":{"summary":"Datasets","operationId":"datasets_api_datasets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]},"post":{"summary":"Upload","operationId":"upload_api_datasets_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_api_datasets_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/datasets/sample":{"post":{"summary":"Sample","operationId":"sample_api_datasets_sample_post","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/datasets/{dataset_id}":{"get":{"summary":"Dataset","operationId":"dataset_api_datasets__dataset_id__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/datasets/{dataset_id}/eda":{"get":{"summary":"Dashboard","operationId":"dashboard_api_datasets__dataset_id__eda_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dataset Id"}},{"name":"target","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/datasets/{dataset_id}/relationship":{"get":{"summary":"Relationship","operationId":"relationship_api_datasets__dataset_id__relationship_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dataset Id"}},{"name":"x","in":"query","required":true,"schema":{"type":"string","title":"X"}},{"name":"y","in":"query","required":true,"schema":{"type":"string","title":"Y"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyses":{"get":{"summary":"Listing","operationId":"listing_api_analyses_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]},"post":{"summary":"Start","operationId":"start_api_analyses_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalysisRequest"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/datasets/{dataset_id}/validate":{"post":{"summary":"Validate Problem","operationId":"validate_problem_api_datasets__dataset_id__validate_post","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dataset Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyses/{analysis_id}":{"get":{"summary":"Get Analysis","operationId":"get_analysis_api_analyses__analysis_id__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"analysis_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Analysis Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyses/{analysis_id}/events":{"get":{"summary":"Events","operationId":"events_api_analyses__analysis_id__events_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"analysis_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Analysis Id"}},{"name":"after","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"After"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyses/{analysis_id}/cancel":{"post":{"summary":"Cancel","operationId":"cancel_api_analyses__analysis_id__cancel_post","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"analysis_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Analysis Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyses/{analysis_id}/retry":{"post":{"summary":"Retry","operationId":"retry_api_analyses__analysis_id__retry_post","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"analysis_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Analysis Id"}}],"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyses/{analysis_id}/artifacts/{kind}":{"get":{"summary":"Artifact","operationId":"artifact_api_analyses__analysis_id__artifacts__kind__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"analysis_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Analysis Id"}},{"name":"kind","in":"path","required":true,"schema":{"type":"string","title":"Kind"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyses/{analysis_id}/predict":{"post":{"summary":"Predict","operationId":"predict_api_analyses__analysis_id__predict_post","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"analysis_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Analysis Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyses/{analysis_id}/chat":{"get":{"summary":"Chat History","operationId":"chat_history_api_analyses__analysis_id__chat_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"analysis_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Analysis Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Chat","operationId":"chat_api_analyses__analysis_id__chat_post","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"analysis_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Analysis Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnalysisRequest":{"properties":{"dataset_id":{"type":"string","format":"uuid","title":"Dataset Id"},"goal":{"type":"string","maxLength":4000,"minLength":5,"title":"Goal"},"task":{"type":"string","enum":["auto","classification","regression","clustering"],"title":"Task","default":"auto"},"target":{"anyOf":[{"type":"string","maxLength":200},{"type":"null"}],"title":"Target"},"features":{"anyOf":[{"items":{"type":"string"},"type":"array","maxItems":100},{"type":"null"}],"title":"Features"},"include_dl":{"type":"boolean","title":"Include Dl","default":true},"use_llm":{"type":"boolean","title":"Use Llm","default":false},"seed":{"type":"integer","maximum":999999.0,"minimum":1.0,"title":"Seed","default":42},"cv_folds":{"type":"integer","maximum":5.0,"minimum":2.0,"title":"Cv Folds","default":3},"tuning_iterations":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Tuning Iterations","default":3},"time_budget_seconds":{"type":"integer","maximum":7200.0,"minimum":10.0,"title":"Time Budget Seconds","default":600},"quick":{"type":"boolean","title":"Quick","default":false}},"additionalProperties":false,"type":"object","required":["dataset_id","goal"],"title":"AnalysisRequest"},"Body_upload_api_datasets_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_api_datasets_post"},"ChatRequest":{"properties":{"question":{"type":"string","maxLength":2000,"minLength":2,"title":"Question"},"use_llm":{"type":"boolean","title":"Use Llm","default":false},"provider":{"anyOf":[{"type":"string","enum":["groq","configured","local"]},{"type":"null"}],"title":"Provider"}},"additionalProperties":false,"type":"object","required":["question"],"title":"ChatRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PredictionRequest":{"properties":{"rows":{"items":{"additionalProperties":{"anyOf":[{"type":"string"},{"type":"number"},{"type":"integer"},{"type":"boolean"},{"type":"null"}]},"type":"object"},"type":"array","maxItems":200,"minItems":1,"title":"Rows"}},"additionalProperties":false,"type":"object","required":["rows"],"title":"PredictionRequest"},"ProblemRequest":{"properties":{"goal":{"type":"string","maxLength":4000,"title":"Goal","default":""},"task":{"type":"string","enum":["auto","classification","regression","clustering"],"title":"Task","default":"auto"},"target":{"anyOf":[{"type":"string","maxLength":200},{"type":"null"}],"title":"Target"},"cv_folds":{"type":"integer","maximum":5.0,"minimum":2.0,"title":"Cv Folds","default":3},"seed":{"type":"integer","maximum":999999.0,"minimum":1.0,"title":"Seed","default":42}},"additionalProperties":false,"type":"object","title":"ProblemRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-API-Key"}}}}